change factorial variable type from long to int for consistency

+++ ConditionalandLoopBasedPrograms/Factorial.java
@@ -8,7 +8,7 @@ public class Factorial {
         System.out.print("Enter a number: ");
         int n = sc.nextInt();
 
-        long fact = 1;
+        int fact = 1;
 
         for (int i = 1; i <= n; i++) {
             fact = fact * i;
